third_party/WebKit/LayoutTests/fast/events/select-event-on-input-recursive.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<input id="input" value="abcd" />
<script>
    window.jsTestIsAsync = true;

    var eventCalledCount = 0;
    input.addEventListener("select", function () {
        eventCalledCount++;
        input.selectionStart = 1;
        input.selectionEnd = 3;
    });

    var animationFrameCount = 0;
    var animation = function () {
        animationFrameCount++;
        shouldBeTrue("animationFrameCount == eventCalledCount");
        if (animationFrameCount == 10)
            window.finishJSTest();
        else
            window.requestAnimationFrame(animation);
    }
    window.requestAnimationFrame(animation);

    input.select();
</script>
</body>
</html>
