third_party/WebKit/LayoutTests/fast/events/select-event-on-input-recursive-expected.txt

PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS animationFrameCount == eventCalledCount is true
PASS successfullyParsed is true

TEST COMPLETE
